feat: Add confirmation dialog before deleting an employee
This commit adds a confirmation dialog before deleting an employee. When the user clicks the delete button, a confirmation dialog is displayed using the SweetAlert2 library. If the user confirms the deletion, the employee is deleted and the employee list is updated. This change improves the user experience by preventing accidental deletions and providing feedback to the user.

+++ src/pages/employee/employee.tsx
@@ -4,6 +4,7 @@ import axios from "axios";
 import { getBaseUrl } from "@/helpers/api";
 import { useEffect, useState } from "react";
 import { EmployeeProps } from "@/types/user";
+import Swal from "sweetalert2";
 
 const Employee = () => {
   const [employees, setEmployees] = useState<EmployeeProps[]>([]);
@@ -45,7 +46,20 @@ const Employee = () => {
   };
 
   const handleDeleteEmployee = (id: number) => {
-    deleteEmployee(id);
+    Swal.fire({
+      title: "Apakah anda yakin?",
+      text: "Anda tidak dapat mengembalikan data yang sudah dihapus",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Ya, hapus",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        deleteEmployee(id);
+        getEmployees();
+      }
+    });
   };
 
   const deleteEmployee = (id: number) => {
